pyproject.toml
[project]
name = "trans_sent"
version = "0.1.0"
dependencies = [
    "annotated-types==0.7.0",
    "blis==1.3.0",
    "catalogue==2.0.10",
    "certifi==2025.8.3",
    "charset-normalizer==3.4.3",
    "click==8.2.1",
    "cloudpathlib==0.21.1",
    "colorama==0.4.6",
    "confection==0.1.5",
    "contourpy==1.3.3",
    "cycler==0.12.1",
    "cymem==2.0.11",
    "filelock==3.19.1",
    "fonttools==4.59.1",
    "fsspec==2025.7.0",
    "huggingface-hub==0.34.4",
    "idna==3.10",
    "jinja2==3.1.6",
    "kiwisolver==1.4.9",
    "langcodes==3.5.0",
    "language-data==1.3.0",
    "marisa-trie==1.2.1",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.2",
    "matplotlib==3.10.5",
    "mdurl==0.1.2",
    "mpmath==1.3.0",
    "murmurhash==1.0.13",
    "networkx==3.3",
    "numpy==2.3.2",
    "packaging==25.0",
    "pandas==2.3.1",
    "pillow==11.3.0",
    "preshed==3.0.10",
    "psutil",
    "pyarrow==21.0.0",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pygments==2.19.2",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "regex==2025.7.34",
    "requests==2.32.4",
    "rich==14.1.0",
    "safetensors==0.6.2",
    "sentencepiece==0.2.1",
    "setuptools==80.9.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "smart-open==7.3.0.post1",
    "spacy==3.8.7",
    "spacy-legacy==3.0.12",
    "spacy-loggers==1.0.5",
    "srsly==2.5.1",
    "sympy==1.13.1",
    "tabulate==0.9.0",
    "thinc==8.3.6",
    "tokenizers==0.21.4",
    "torch==2.5.1+cu121",
    "torchaudio==2.5.1+cu121",
    "torchvision==0.20.1+cu121",
    "tqdm==4.67.1",
    "transformers==4.55.2",
    "typer==0.16.0",
    "typing-extensions==4.14.1",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "wasabi==1.1.3",
    "weasel==0.4.1",
    "wrapt==1.17.3",
]
